Adds tests for ActionConfig.create()
During the process, I've moved the validation
from `Config` to each specific action. Currently
this only affects year and ranges, whose create
can return an error.

Because of this, the anonymisations function can
now fail and returns an error in case there name
can't be match or if the specific action can't be
created.

Signed-off-by: Albert Pastrana <albert.pastrana@gmail.com>